Input validator works
